chore: fix send message
