+++ src/reducer/conversationSlice.action.js
@@ -4,7 +4,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { Platform } from 'react-native';
 
 import { createPendingMessage, buildCreatePayload } from 'helpers/conversationHelpers';
-import { addOrUpdateMessageHelper } from './conversationHelpers';
+import { addOrUpdateMessage } from 'reducer/conversationSlice';
 
 import axios from 'helpers/APIHelper';
 import * as RootNavigation from 'helpers/NavigationHelper';
@@ -63,6 +63,7 @@ const actions = {
         const {
           conversations: { currentInbox, assigneeType, conversationStatus },
         } = getState();
+        // TODO: Move to helpers since this is used in multiple places
         const params = {
           inbox_id: currentInbox || null,
           assignee_type: assigneeType === 'mine' ? 'me' : assigneeType,
@@ -95,6 +96,7 @@ const actions = {
         const {
           data: { payload, meta },
         } = response;
+        // TODO: Commit meta to store
         return { data: payload, meta, conversationId };
       } catch (error) {
         if (!error.response) {
@@ -155,21 +157,20 @@ const actions = {
   ),
   sendMessage: createAsyncThunk(
     'conversations/sendMessage',
-    async ({ data }, { dispatch, getState, rejectWithValue }) => {
+    async ({ data }, { dispatch, rejectWithValue }) => {
       const { conversationId } = data;
       let payload;
       const pendingMessage = createPendingMessage(data);
 
       try {
-        const state = getState();
-        const newState = { ...state.conversations };
-        addOrUpdateMessageHelper(newState, {
-          ...pendingMessage,
-          status: MESSAGE_STATUS.PROGRESS,
-        });
-        dispatch({ type: 'conversations/addOrUpdateMessage', payload: newState });
-
-        payload = buildCreatePayload(data);
+        dispatch(
+          addOrUpdateMessage({
+            ...pendingMessage,
+            status: MESSAGE_STATUS.PROGRESS,
+          }),
+        );
+        payload = buildCreatePayload(pendingMessage);
+        const { file } = data;
 
         const contentType =
           Platform.OS === 'ios' && file
@@ -183,21 +184,19 @@ const actions = {
             'Content-Type': contentType,
           },
         });
-
-        const finalState = { ...getState().conversations };
-        addOrUpdateMessageHelper(finalState, {
-          ...response.data,
-          status: MESSAGE_STATUS.SENT,
-        });
-        dispatch({ type: 'conversations/addOrUpdateMessage', payload: finalState });
-        return response.data;
+        dispatch(
+          addOrUpdateMessage({
+            ...response.data,
+            status: MESSAGE_STATUS.SENT,
+          }),
+        );
       } catch (error) {
-        const errorState = { ...getState().conversations };
-        addOrUpdateMessageHelper(errorState, {
-          ...pendingMessage,
-          status: MESSAGE_STATUS.FAILED,
-        });
-        dispatch({ type: 'conversations/addOrUpdateMessage', payload: errorState });
+        dispatch(
+          addOrUpdateMessage({
+            ...pendingMessage,
+            status: MESSAGE_STATUS.FAILED,
+          }),
+        );
         if (!error.response) {
           throw error;
         }
@@ -205,6 +204,7 @@ const actions = {
       }
     },
   ),
+
   toggleConversationStatus: createAsyncThunk(
     'conversations/toggleConversationStatus',
     async ({ conversationId, status, snoozedUntil = null }, { rejectWithValue }) => {
@@ -269,6 +269,7 @@ const actions = {
       try {
         const apiUrl = `conversations/${conversationId}/assignments?assignee_id=${assigneeId}`;
         const response = await axios.post(apiUrl);
+
         return response.data;
       } catch (error) {
         if (!error.response) {
@@ -282,6 +283,7 @@ const actions = {
     'conversations/toggleTypingStatus',
     async ({ conversationId, typingStatus }, { rejectWithValue }) => {
       const apiUrl = `conversations/${conversationId}/toggle_typing_status`;
+
       await axios
         .post(apiUrl, {
           typing_status: typingStatus,
